Ensure test-image-openj9 target fails when java -version fails
Previously, the exit status could come from tee, which succeeds.

There is logic in make/autoconf/basics.m4 that checks for support and
then opportunistically adds -o pipefail to SHELL in the makefiles, which
already accomplishes the same thing, but only when pipefail support is
detected. That should be most of the time, but it's not obvious that we
should expect pipefail support in all cases.

With this change, the target will fail whenever java -version fails,
regardless of the pipefail setting.